# docs/product-service-api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Product Service API",
    "description": "API каталога умных устройств для дома",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://192.168.1.115:52700",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Товары",
      "description": "Управление каталогом умных устройств для дома"
    },
    {
      "name": "Категории",
      "description": "Управление категориями товаров"
    }
  ],
  "paths": {
    "/api/products": {
      "get": {
        "tags": [
          "Товары"
        ],
        "summary": "Получить все товары",
        "description": "Возвращает список всех активных товаров",
        "operationId": "getAllProducts",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Товары"
        ],
        "summary": "Создать товар",
        "description": "Добавляет новый товар в каталог",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": [
          "Категории"
        ],
        "summary": "Получить все категории",
        "description": "Возвращает список всех категорий товаров",
        "operationId": "getAllCategories",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Категории"
        ],
        "summary": "Создать категорию",
        "description": "Создаёт новую категорию товаров",
        "operationId": "createCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Товары"
        ],
        "summary": "Получить товар по ID",
        "description": "Возвращает подробную информацию о товаре",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID товара",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Товары"
        ],
        "summary": "Обновить товар",
        "description": "Частичное обновление информации о товаре",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID товара",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/search": {
      "get": {
        "tags": [
          "Товары"
        ],
        "summary": "Поиск товаров",
        "description": "Поиск товаров по названию (частичное совпадение)",
        "operationId": "searchProducts",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "Поисковый запрос",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/category/{categoryId}": {
      "get": {
        "tags": [
          "Товары"
        ],
        "summary": "Товары по категории",
        "description": "Возвращает товары указанной категории",
        "operationId": "getProductsByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "ID категории",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "tags": [
          "Категории"
        ],
        "summary": "Получить категорию по ID",
        "operationId": "getCategoryById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP-статус",
            "format": "int32",
            "example": 404
          },
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Товар не найден"
          },
          "timestamp": {
            "type": "string",
            "description": "Время ошибки",
            "format": "date-time"
          },
          "validationErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "description": "Детали ошибок валидации (если есть)"
            },
            "description": "Детали ошибок валидации (если есть)"
          }
        },
        "description": "Ответ с информацией об ошибке"
      },
      "CreateProductRequest": {
        "required": [
          "name",
          "price"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "description": "Название товара",
            "example": "Умная лампа ShtLight LED"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string",
            "description": "Описание товара",
            "example": "Светодиодная лампа с поддержкой Wi-Fi"
          },
          "price": {
            "minimum": 0.01,
            "exclusiveMinimum": false,
            "type": "number",
            "description": "Цена в рублях",
            "example": 2490.0
          },
          "categoryId": {
            "type": "integer",
            "description": "ID категории",
            "format": "int64",
            "example": 1
          },
          "imageUrl": {
            "type": "string",
            "description": "URL изображения товара"
          }
        },
        "description": "Запрос на создание товара"
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID категории",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Название категории",
            "example": "Освещение"
          },
          "description": {
            "type": "string",
            "description": "Описание категории",
            "example": "Умные лампы и светильники"
          }
        },
        "description": "Категория товаров"
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID товара",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Название товара",
            "example": "Умная лампа ShtLight LED"
          },
          "description": {
            "type": "string",
            "description": "Описание товара",
            "example": "Светодиодная лампа с поддержкой Wi-Fi"
          },
          "price": {
            "type": "number",
            "description": "Цена в рублях",
            "example": 2490.0
          },
          "category": {
            "$ref": "#/components/schemas/CategoryDto"
          },
          "imageUrl": {
            "type": "string",
            "description": "URL изображения товара"
          },
          "active": {
            "type": "boolean",
            "description": "Товар активен и доступен для продажи",
            "example": true
          }
        },
        "description": "Информация о товаре"
      },
      "CreateCategoryRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "description": "Название категории",
            "example": "Освещение"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "description": "Описание категории",
            "example": "Умные лампы и светильники"
          }
        },
        "description": "Запрос на создание категории"
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "description": "Название товара",
            "example": "Умная лампа ShtLight LED v2"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string",
            "description": "Описание товара"
          },
          "price": {
            "minimum": 0.01,
            "exclusiveMinimum": false,
            "type": "number",
            "description": "Цена в рублях",
            "example": 2990.0
          },
          "categoryId": {
            "type": "integer",
            "description": "ID категории",
            "format": "int64",
            "example": 1
          },
          "imageUrl": {
            "type": "string",
            "description": "URL изображения товара"
          },
          "active": {
            "type": "boolean",
            "description": "Товар активен",
            "example": true
          }
        },
        "description": "Запрос на обновление товара"
      }
    }
  }
}